tesseract/text_extraction.py
import pytesseract
from PIL import Image

pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'

def extract_text_from_image(image):
    """이미지에서 텍스트 추출"""
    return pytesseract.image_to_string(image)