requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
langdetect==1.0.9
MouseInfo==0.1.3
packaging==24.0
pillow==10.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.31.0
six==1.16.0
urllib3==2.2.1
